+++ src/lib/theme/Header.svelte
@@ -116,7 +116,7 @@ afterNavigate(() => {
     </nav>
      <?php get_search_form(['a309_menu' => true]); ?>  -->
 
-<button on:click={themeSwitch} aria-label="Theme Switch" class="ml-5 w-16 h-6 rounded-full bg-white flex items-center transition duration-300 focus:outline-none shadow">
+<button on:click={themeSwitch} aria-label="Theme Switch" class="{`ml-5 w-16 h-6 rounded-ful flex items-center transition duration-300 focus:outline-none shadow ${theme === 'dark' ? 'bg-gray-700': 'bg-white'}`}">
   <div bind:this={svgContainer} class="{`border border-white w-9 h-9 relative rounded-full transition duration-500 transform ${theme === 'dark' ? 'bg-gray-700 translate-x-full' : 'bg-yellow-500 -translate-x-2' } p-1 text-white`}">
       <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor">
       <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 3v1m0 16v1m9-9h-1M4 12H3m15.364 6.364l-.707-.707M6.343 6.343l-.707-.707m12.728 0l-.707.707M6.343 17.657l-.707.707M16 12a4 4 0 11-8 0 4 4 0 018 0z" />