+++ src/lib/utils/server/widgets.ts
@@ -55,7 +55,9 @@ export const getTagCloud = async () => {
                 value: tagCloud,
                 updated_at: new Date().toISOString()
             }
-        )
+        ).then((res) => {
+            console.log("Operation result: ", res)
+        })
         return tagCloud
     } else {
         const date = new Date(data.updated_at)
@@ -64,7 +66,9 @@ export const getTagCloud = async () => {
         const diffDays = Math.ceil(diff / (1000 * 60 * 60 * 24))
         if(diffDays > 3) {
                 generateTagCloud().then((res) => {
-                supabase.from('fsk_blog_store').upsert( { name: 'tagCloud', value: res, updated_at: new Date().toISOString() })
+                supabase.from('fsk_blog_store').upsert( { name: 'tagCloud', value: res, updated_at: new Date().toISOString() }).then((res) => {
+                    console.log("Operation result: ", res)
+                    })
                 })
                 
             }
@@ -107,9 +111,10 @@ export const getGoodReadsData = async () => {
         const diffDays = Math.ceil(diff / (1000 * 60 * 60 * 24))
         if(diffDays > 7) {
                 createGoodReadsData().then((res) => {
-                supabase.from('fsk_blog_store').upsert( { name: 'goodreads', value: res, updated_at: new Date().toISOString() })
+                supabase.from('fsk_blog_store').upsert( { name: 'goodreads', value: res, updated_at: new Date().toISOString() }).then((res) => {
+                    console.log("Operation result: ", res)
                 })
-                
+            })
             }
         }
         return data.value