+++ src/routes/(index)/author/[slug]/+page.server.ts
@@ -8,17 +8,22 @@ import { supabase } from '@/lib/node/supaClientFS'
 import { error } from '@/lib/utils/page'
 import { config } from '$lib/config'
 
+const defaultError = {
+  message: 'Author not found',
+  pageType: 'author'
+} as App.Error
+
 export const load: PageServerLoad = async (req) => {
     const slug  = req.params.slug
 
     if(!slug) {
-        throw error(404, 'Not found')
+        throw error(404, defaultError)
     }
     const user = (await supabase.from('fsk_blog_author')
         .select('*')
         .eq('username', slug).single()).data
         if(!user) {
-            throw error(404, 'Not found')
+            throw error(404, defaultError)
         }
     
     const fallBackDescription =  truncate(user.description, 160, {
@@ -26,7 +31,7 @@ export const load: PageServerLoad = async (req) => {
     }) || 'Desciption is missing | ' + config.siteName 
     return appendToData(
       {res: user}, 
-      {pageTitle: 'Author | ' + user.username ?? 'Author | ' + config.siteName, 
+      {pageTitle: user?.username ? 'Author | ' + user.username : 'Author | ' + config.siteName, 
       pageDescription: fallBackDescription    
     })
   }