+++ src/routes/(index)/preview/+page.server.ts
@@ -4,6 +4,11 @@ import { checkData, appendToData } from '@/lib/utils/page'
 import { error } from '@/lib/utils/page'
 import { supabase } from '@/lib/node/supaClientFS'
 
+const defaultError = {
+    message: 'Unauthorized',
+    pageType: 'author'
+    } as App.Error
+
 export const load: PageServerLoad = async (req) => {
     const username = req.cookies.get('username')
     const hash = req.cookies.get('hash')
@@ -14,7 +19,7 @@ export const load: PageServerLoad = async (req) => {
         .eq('username', username).single()).data
         if(user) {
             if(hash !== user.password_hash) {
-                throw error(401, 'Unauthorized')
+                throw error(401, defaultError)
             }
         }
     }