+++ src/routes/(index)/tag/[slug]/[[page=page]]/[[num=num]]/+page.svelte
@@ -4,6 +4,8 @@
     import SideBar from '@/lib/theme/SideBar.svelte';
     import { isLoading } from '@/stores/main';
     import { config } from '$lib/config';
+    import { fcFrame, getDefaultButtons } from '$lib/utils/server/fc-frame'
+
 
     export let data: any;
     const sidebarData = data?.sidebarData
@@ -22,7 +24,13 @@
 <meta property="og:description" content="{data.pageDescription}">
 <meta property="og:type" content="website" />
 <meta property="og:url" content={`${$SveltePage.url}`} />
-<meta property="og:image" content="https://flashsoft.eu/res/flashsoftLogo.png" />
+<meta property="og:image" content={`${config.baseSiteUrl}/images/og/default-og-blog-opt.webp`} />
+{`${fcFrame({
+    image: `${config.baseSiteUrl}/images/og/default-og-blog-opt.webp`,
+    postUrl: `${config.baseSiteUrl}/fc-frame-handler`,
+    buttons: getDefaultButtons($SveltePage.url.toString())
+})}`}
+
 <link rel="alternate" type="application/rss+xml" title="{`${config.siteName} » Feed`}" href="{`${config.baseSiteUrl}/feed`}">
 <link rel="canonical" href={`${$SveltePage.url}`} />